Prevent null lockscreen credential
The getPasswordHistoryHashFactor API expects a non-null credential
parameter. If the existing credential was previously set to none, the
credential will be null. Therefore, it is necessary to check for this
null state and replace it with a "none" lockscreen credential.

Bug: 230576954
Test: manual
Change-Id: Ia99547c1c93aed41471be1c0f074458e8d3de9ae
(cherry picked from commit 1560c4dc7ee083a1d50953b46bc9662ea759d517)